agent/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
claude-agent-sdk>=0.1.0
sse-starlette>=2.0.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.19.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
